+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Changed
 - Fixed `EmptyStackException` when writing a new entry type to the very first release
 
+### Removed
+- Redundant empty line when inserting a subheader above an existing subheader
+
 ## [0.1.0] - 2022-05-19
 ### Added
 - Initial version. Adds PR title as changelog entry